Deploy script fixes

+++ deploy_win64.ps1
@@ -1,15 +1,26 @@
 git pull
+
+# test
+if($test)
+{
+	$testResult = dotnet test
+	Write-Output "Running tests"
+	if($testResult.contains("Failed: "))
+	{
+		Write-Error $testResult
+		return -1
+	}
+	Write-Output "All tests passed"
+}
+
 # Kill running processes
 $p = Get-Process -Name "DodoServer"
-Stop-Process -InputObject $p
-Get-Process | Where-Object {$_.HasExited}
+if($p)
+{
+	Stop-Process -InputObject $p
+	Get-Process | Where-Object {$_.HasExited}
+}
 
-$p = Get-Process -Name "node"
-Stop-Process -InputObject $p
-Get-Process | Where-Object {$_.HasExited}
-
-# test - TODO
-# dotnet test --project test\Dodo.UnitTests\Dodo.UnitTests.csproj
 # publish
 dotnet publish --force src\DodoServer\DodoServer.csproj -o ..\build -c Debug
